--- a/src/Pact/Native.hs
+++ b/src/Pact/Native.hs
@@ -117,6 +117,7 @@ import Pact.Types.Purity
 import Pact.Types.Runtime
 import Pact.Types.Version
 import Pact.Types.Namespace
+import Crypto.Hash.Keccak256Native (Keccak256Error(..), keccak256)
 import Crypto.Hash.PoseidonNative (poseidon)
 import Crypto.Hash.HyperlaneMessageId (hyperlaneMessageId)
 
@@ -136,10 +137,9 @@ natives =
   , decryptDefs
   , guardDefs
   , zkDefs
-  , poseidonHackAChainDefs
+  , hashDefs
   , hyperlaneDefs
   ]
-
 
 -- | Production native modules as a dispatch map.
 nativeDefs :: HM.HashMap Text Ref
@@ -147,7 +147,6 @@ nativeDefs = mconcat $ map moduleToMap natives
 
 moduleToMap :: NativeModule -> HM.HashMap Text Ref
 moduleToMap = HM.fromList . map (asString *** Direct) . snd
-
 
 lengthDef :: NativeDef
 lengthDef = defRNative "length" length' (funType tTyInteger [("x",listA)])
@@ -1564,8 +1563,11 @@ base64DecodeWithShimmedErrors i txt = do
       Nothing ->
         evalError i "Could not parse error message"
 
-poseidonHackAChainDefs :: NativeModule
-poseidonHackAChainDefs = ("Poseidon Hash", [ poseidonHackAChainDef ])
+hashDefs :: NativeModule
+hashDefs = ("Hashes",)
+  [ poseidonHackAChainDef
+  , keccak256Def
+  ]
 
 poseidonHackAChainDef :: NativeDef
 poseidonHackAChainDef = defGasRNative
@@ -1586,6 +1588,47 @@ poseidonHackAChainDef = defGasRNative
       = computeGas' i (GPoseidonHashHackAChain $ length as) $
         return $ toTerm $ poseidon intArgs
      | otherwise = argsError i as
+
+keccak256Def :: NativeDef
+keccak256Def = defGasRNative
+  "hash-keccak256"
+  keccak256'
+  (funType tTyString [("bytes", TyList tTyString)])
+  [
+    "(hash-keccak256 [])"
+  , "(hash-keccak256 [\"\"])"
+  , "(hash-keccak256 [\"T73FllCNJKKgAQ4UCYC4CfucbVXsdRJYkd2YXTdmW9gPm-tqUCB1iKvzzu6Md82KWtSKngqgdO04hzg2JJbS-yyHVDuzNJ6mSZfOPntCTqktEi9X27CFWoAwWEN_4Ir7DItecXm5BEu_TYGnFjsxOeMIiLU2sPlX7_macWL0ylqnVqSpgt-tvzHvJVCDxLXGwbmaEH19Ov_9uJFHwsxMmiZD9Hjl4tOTrqN7THy0tel9rc8WtrUKrg87VJ7OR3Rtts5vZ91EBs1OdVldUQPRP536eTcpJNMo-N0fy-taji6L9Mdt4I4_xGqgIfmJxJMpx6ysWmiFVte8vLKl1L5p0yhOnEDsSDjuhZISDOIKC2NeytqoT9VpBQn1T3fjWkF8WEZIvJg5uXTge_qwA46QKV0LE5AlMKgw0cK91T8fnJ-u1Dyk7tCo3XYbx-292iiih8YM1Cr1-cdY5cclAjHAmlglY2ia_GXit5p6K2ggBmd1LpEBdG8DGE4jmeTtiDXLjprpDilq8iCuI0JZ_gvQvMYPekpf8_cMXtTenIxRmhDpYvZzyCxek1F4aoo7_VcAMYV71Mh_T8ox7U1Q4U8hB9oCy1BYcAt06iQai0HXhGFljxsrkL_YSkwsnWVDhhqzxWRRdX3PubpgMzSI290C1gG0Gq4xfKdHTrbm3Q\"])"
+  , "(hash-keccak256 [\"T73FllCNJKKgAQ4UCYC4CfucbVXsdRJYkd2YXTdmW9g\", \"D5vralAgdYir887ujHfNilrUip4KoHTtOIc4NiSW0vs\", \"LIdUO7M0nqZJl84-e0JOqS0SL1fbsIVagDBYQ3_givs\", \"DItecXm5BEu_TYGnFjsxOeMIiLU2sPlX7_macWL0ylo\", \"p1akqYLfrb8x7yVQg8S1xsG5mhB9fTr__biRR8LMTJo\", \"JkP0eOXi05Ouo3tMfLS16X2tzxa2tQquDztUns5HdG0\", \"ts5vZ91EBs1OdVldUQPRP536eTcpJNMo-N0fy-taji4\", \"i_THbeCOP8RqoCH5icSTKcesrFpohVbXvLyypdS-adM\", \"KE6cQOxIOO6FkhIM4goLY17K2qhP1WkFCfVPd-NaQXw\", \"WEZIvJg5uXTge_qwA46QKV0LE5AlMKgw0cK91T8fnJ8\", \"rtQ8pO7QqN12G8ftvdoooofGDNQq9fnHWOXHJQIxwJo\", \"WCVjaJr8ZeK3mnoraCAGZ3UukQF0bwMYTiOZ5O2INcs\", \"jprpDilq8iCuI0JZ_gvQvMYPekpf8_cMXtTenIxRmhA\", \"6WL2c8gsXpNReGqKO_1XADGFe9TIf0_KMe1NUOFPIQc\", \"2gLLUFhwC3TqJBqLQdeEYWWPGyuQv9hKTCydZUOGGrM\", \"xWRRdX3PubpgMzSI290C1gG0Gq4xfKdHTrbm3Q\"])"
+  ]
+  "Compute the hash of a list of unpadded base64url-encoded inputs. The hash is computed incrementally over all of the decoded inputs."
+  where
+    keccak256' :: RNativeFun e
+    keccak256' i = \case
+      args@[TList ls _ _] -> do
+        texts <- forM ls $ \case
+          TLiteral (LString s) _ -> pure s
+          _ -> argsError i args
+
+        -- At first I thought we should charge based on decoded bytes.
+        -- Both base64 and base64url have predictable byte-lengths from
+        -- both the input and output. However, charging on the encoded
+        -- length makes more sense, for two reasons:
+        --
+        --   1. We do not know if the decoding will fail here, and that
+        --      decoding is best left to the impl, not the nativedef
+        --   2. Using the encoded bytes will potentially overcharge a bit,
+        --      but that's better than undercharging
+        let chunkBytes = V.map (BS.length . T.encodeUtf8) texts
+        computeGas' i (GKeccak256 chunkBytes) (pure ())
+
+        output <- case keccak256 texts of
+          Left (Keccak256Base64Exception msg) -> evalError' i ("Base64URL decode failed: " <> prettyString msg)
+          Left (Keccak256OpenSslException msg) -> evalError' i ("OpenSSL error when keccak256 hashing: " <> prettyString msg)
+          Left (Keccak256OtherException exc) -> evalError' i ("Exception when keccak256 hashing: " <> prettyString (displayException exc))
+          Right output -> pure output
+        pure (toTerm output)
+
+      args -> argsError i args
 
 hyperlaneDefs :: NativeModule
 hyperlaneDefs = ("Hyperlane",)
@@ -1653,7 +1696,7 @@ hyperlaneDecodeTokenMessageDef =
                 -- the Binary library, and we will suppress it to shield ourselves
                 -- from forking behavior if we update our Binary version.
                 Left (_,_,e) | "TokenMessage" `isPrefixOf` e -> evalError' i $ "Decoding error: " <> pretty e
-                Left _ -> evalError' i "Decoding error: binary decoding failed" 
+                Left _ -> evalError' i "Decoding error: binary decoding failed"
                 Right (_,_,(amount, chain, recipient)) ->
                   case PGuard <$> J.eitherDecode (BS.fromStrict  $ T.encodeUtf8 recipient) of
                     Left _ -> evalError' i $ "Could not parse recipient into a guard"