--- a/src/Pact/Types/Gas.hs
+++ b/src/Pact/Types/Gas.hs
@@ -189,6 +189,8 @@ data GasArgs
   | GHyperlaneDecodeTokenMessage !Int
   -- ^ Cost of hyperlane-decode-token-message on this size (in bytes) of the
   --   hyperlane TokenMessage base64-encoded string.
+  | GKeccak256 !(V.Vector Int)
+  -- ^ Cost of hash-keccak256 given the number of bytes in each chunk.
 
 data IntOpThreshold
   = Pact43IntThreshold
@@ -259,6 +261,7 @@ instance Pretty GasArgs where
     GPoseidonHashHackAChain len -> "GPoseidonHashHackAChain:" <> pretty len
     GHyperlaneMessageId len -> "GHyperlaneMessageId:" <> pretty len
     GHyperlaneDecodeTokenMessage len -> "GHyperlaneDecodeTokenMessage:" <> pretty len
+    GKeccak256 chunksBytes -> "GKeccak256:" <> pretty (V.toList chunksBytes)
 
 newtype GasLimit = GasLimit ParsedInteger
   deriving (Eq,Ord,Generic)