--- a/tests/AnalyzeSpec.hs
+++ b/tests/AnalyzeSpec.hs
@@ -2814,6 +2814,23 @@ spec = describe "analyze" $ do
           |]
     expectPass code $ Valid Success'
 
+  describe "keccak256" $ do
+    let code =
+          [text|
+            (defun test:bool (arg:[string])
+              (enforce (=
+                (hash-keccak256 [""])
+                "xdJGAYb3IzySfn2y3McDwOUAtlPKgic7e_rYBF2FpHA") "should match empty")
+
+               (enforce (=
+                 (hash-keccak256 [""])
+                 (hash-keccak256 arg)) "should match empty")
+
+
+            )
+          |]
+    expectPass code $ Valid Success'
+
   describe "validate-principal" $ do
     let code =
          [text| (defun test:bool (ks: keyset)