--- a/tests/PactTests.hs
+++ b/tests/PactTests.hs
@@ -25,6 +25,7 @@ import qualified GasModelSpec
 import qualified GoldenSpec
 import qualified HistoryServiceSpec
 import qualified HyperlaneSpec
+import qualified Keccak256Spec
 import qualified PactContinuationSpec
 import qualified PersistSpec
 import qualified RemoteVerifySpec
@@ -62,6 +63,7 @@ main = hspec $ parallel $ do
   describe "GoldenSpec" GoldenSpec.spec
   describe "HistoryServiceSpec" HistoryServiceSpec.spec
   describe "HyperlaneSpec" HyperlaneSpec.spec
+  describe "Keccak256Spec" Keccak256Spec.spec
   describe "PactContinuationSpec" PactContinuationSpec.spec
   describe "PersistSpec" PersistSpec.spec
   describe "RemoteVerifySpec" RemoteVerifySpec.spec