… docs/en/pact-functions.md
… golden/gas-model/golden
… tests/Keccak256Spec.hs