@chessai 2024-03-20
> Cannot be merged until we remove the unsafe lens access

I thought the lens access was safe because I was under the impression that natives were typechecked, but I was wrong